requirements.txt
python-twitter==3.1
Scrapy==0.24.4
